Added the time counter in make_checkpoint

+++ eo/src/do/make_continue.h
@@ -94,7 +94,7 @@ eoContinue<Indi> & do_make_continue(eoParser& _parser, eoState& _state, eoEvalFu
       }
 
   // Same thing with Eval - but here default value is 0
-  eoValueParam<unsigned>& maxEvalParam = _parser.createParam(unsigned(0), "maxEval", "Maximum number of evaluations (0 = none)",'E',"Stopping criterion");
+  eoValueParam<unsigned long>& maxEvalParam = _parser.createParam((unsigned long)0, "maxEval", "Maximum number of evaluations (0 = none)",'E',"Stopping criterion");
 
     if (maxEvalParam.value()) // positive: -> define and store
       {
